requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
spacy
nltk
transformers
diffusers
torch
torchvision
torchaudio
Pillow
opencv-python
reportlab
pytest
accelerate
safetensors
certifi
